+++ src/dsp/filtering/fir/windows/polynomial.rs
@@ -3,7 +3,7 @@ use super::window::*;
 
 pub struct RectangularWindow {}
 impl WindowFunction for RectangularWindow {
-    fn window_function(&self, sample: f32, window_size: f32) -> f32 {
+    fn window_function(&self, sample: u32, window_size: usize) -> f32 {
         return 1.0;
     }
 }
@@ -11,17 +11,17 @@ impl WindowFunction for RectangularWindow {
 
 pub struct TriangularWindow {}
 impl WindowFunction for TriangularWindow {
-    fn window_function(&self, sample: f32, window_size: f32) -> f32 {
-        return 1.0 - ((sample - (window_size / 2.0)) / (window_size / 2.0)).abs();
+    fn window_function(&self, sample: u32, window_size: usize) -> f32 {
+        return 1.0 - ((sample as f32 - (window_size as f32 / 2.0)) / (window_size as f32 / 2.0)).abs();
     }
 }
 
 
 pub struct WelchWindow {}
 impl WindowFunction for WelchWindow {
-    fn window_function(&self, sample: f32, window_size: f32) -> f32 {
-        let size_by_2 = window_size / 2.0;
-        let inner_term = (sample - size_by_2) / size_by_2;
+    fn window_function(&self, sample: u32, window_size: usize) -> f32 {
+        let size_by_2 = window_size as f32 / 2.0;
+        let inner_term = (sample as f32 - size_by_2) / size_by_2;
         return 1.0 - (inner_term).powf(2.0);
     }
 }
